+++ example/__init__.py
@@ -8,15 +8,21 @@ from pyramid_contextauth import authentication_policy
 
 def includeme(config):
     config.include('pyramid_contextauth')
+    tkt_policy = AuthTktAuthenticationPolicy('secret')
     config.register_authentication_policy(
-        AuthTktAuthenticationPolicy('secret'),
+        tkt_policy,
         Context1,
+    )
+    config.register_authentication_policy(
+        tkt_policy,
+        Context2,
     )
     config.register_authentication_policy(
         BasicAuthAuthenticationPolicy('realm'),
         Context2,
     )
     config.scan()
+    config.commit()
 
 
 class Context1(object):
@@ -35,6 +41,10 @@ class Context4(object):
     pass
 
 
+class Context5(Context4):
+    pass
+
+
 @authentication_policy(Context3, Context4)
-class Context3Policy(object):
+class Context345Policy(object):
     pass