+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Development
+-----------
+
+* Add introspectables for registered authentication policies.
+
 0.3
 ---